Added code
Made code for the ball, the point values of the bumpers and tunnels, and the high score. Still need a way to launch the ball, a limited # of balls, destruction of the ball, and a spawn point for the ball. Also I want to make the bumpers light up when the ball hits them. Might borrow from HW #2 and HW #3 a bit. I think I can add one last component at the very top of the table, past the two arches. Maybe highest point value goal? Also, I still need to figure out where to put Bill. Maybe if you get a certain point score, he rises up behind the table and just stares at the player. That'd be interesting and creepy at the same time. I'll keep it in mind.