Various Additions
Worked on ball code, I think it will destroy the ball when it hits the Rujnuj block, but the points won't work until I have the canvas set up. Need to figure out the spawn point limit, I think that the respawn should be triggered immediately after the destruction of the previous ball, but every respawn subtracts 1 from the ball counter. If the ball counter is zero and the ball is destroyed, it will trigger the Game Over Screen. Once I have all of this at least mostly figured out, I need to work on the ball launcher. I think I'm gonna modify the flipper code and greatly increase the hitting power.