Fixes and Spawn Point
Fixed the High Score (which is stuck at ~1000 so I just increased the point values of the bumbers and tunnels) and fixed the tunnels (weren't giving points when ball passed thru arches, made tiny platforms under each arch that gives points instead of ghost box.) Finally added the spawn point for balls, as well as a limited number. At this point, the game *could* be considered "done", but it's so bare bones that I'm going to keep working on it.